refactor: tidy up and organise server files
