+++ src/server/ddb.js
@@ -8,7 +8,6 @@ const {
   PutCommand,
   DynamoDBDocumentClient,
 } = require("@aws-sdk/lib-dynamodb");
-
 const getAWSCredentialsDependingOnEnvironment =
   require("./utils").getAWSCredentialsDependingOnEnvironment;
 
@@ -26,56 +25,56 @@ const getDynamoDBTableName = async () => {
   return response.Parameter.Value;
 };
 
-const saveControlMessage = async (oscMsg) => {
-  const credentials = getAWSCredentialsDependingOnEnvironment();
-  const tableName = await getDynamoDBTableName();
+module.exports = {
+  saveControlMessage: async (oscMsg) => {
+    const credentials = getAWSCredentialsDependingOnEnvironment();
+    const tableName = await getDynamoDBTableName();
 
-  console.log(
-    `Saving control message ${JSON.stringify(oscMsg)} to DynamoDB table: ${tableName}`,
-  );
+    console.log(
+      `Saving control message ${JSON.stringify(oscMsg)} to DynamoDB table: ${tableName}`,
+    );
 
-  const ddbClient = new DynamoDBClient({
-    region: "eu-west-2",
-    credentials: credentials,
-  });
-  const docClient = DynamoDBDocumentClient.from(ddbClient);
+    const ddbClient = new DynamoDBClient({
+      region: "eu-west-2",
+      credentials: credentials,
+    });
+    const docClient = DynamoDBDocumentClient.from(ddbClient);
 
-  const command = new PutCommand({
-    TableName: tableName,
-    Item: {
-      channelAndGroup: `${oscMsg.args[1][0]}${oscMsg.args[1][1]}`,
-      channel: oscMsg.address,
-      args: oscMsg.args,
-      timestamp: Date.now().toString(),
-    },
-  });
+    const command = new PutCommand({
+      TableName: tableName,
+      Item: {
+        channelAndGroup: `${oscMsg.args[1][0]}${oscMsg.args[1][1]}`,
+        channel: oscMsg.address,
+        args: oscMsg.args,
+        timestamp: Date.now().toString(),
+      },
+    });
 
-  const response = await docClient.send(command);
-  return response;
+    const response = await docClient.send(command);
+    return response;
+  },
+
+  getControlMessages: async () => {
+    const credentials = getAWSCredentialsDependingOnEnvironment();
+    const tableName = await getDynamoDBTableName();
+    const ddbClient = new DynamoDBClient({
+      region: "eu-west-2",
+      credentials: credentials,
+    });
+    const docClient = DynamoDBDocumentClient.from(ddbClient);
+
+    console.log(
+      `Fetching all control messages from DynamoDB table: ${tableName}`,
+    );
+
+    const command = new ScanCommand({
+      TableName: tableName,
+    });
+
+    const response = await docClient.send(command);
+    return response.Items.map((item) => ({
+      address: item.channel,
+      args: ["loader", item.args],
+    }));
+  },
 };
-
-const getControlMessages = async () => {
-  const credentials = getAWSCredentialsDependingOnEnvironment();
-  const tableName = await getDynamoDBTableName();
-  const ddbClient = new DynamoDBClient({
-    region: "eu-west-2",
-    credentials: credentials,
-  });
-  const docClient = DynamoDBDocumentClient.from(ddbClient);
-
-  console.log(
-    `Fetching all control messages from DynamoDB table: ${tableName}`,
-  );
-
-  const command = new ScanCommand({
-    TableName: tableName,
-  });
-
-  const response = await docClient.send(command);
-  return response.Items.map((item) => ({
-    address: item.channel,
-    args: ["loader", item.args],
-  }));
-};
-
-module.exports = { saveControlMessage, getControlMessages };
