+++ src/server/utils.js
@@ -3,23 +3,20 @@ const ipInt = require("ip-to-int");
 const { fromSSO } = require("@aws-sdk/credential-provider-sso");
 const { fromInstanceMetadata } = require("@aws-sdk/credential-providers");
 
-const generateNameForIp = (ip) => {
-  const seed = ipInt(ip).toInt();
-  return uniqueNamesGenerator({
-    dictionaries: [names],
-    seed: seed,
-  }).toLowerCase();
-};
-
-const getAWSCredentialsDependingOnEnvironment = () => {
-  if (process.env.NODE_ENV == "production") {
-    return fromInstanceMetadata()();
-  } else {
-    return fromSSO({ profile: "osc-chat" })();
-  }
-};
-
 module.exports = {
-  generateNameForIp,
-  getAWSCredentialsDependingOnEnvironment,
+  generateNameFromIp: (ip) => {
+    const seed = ipInt(ip).toInt();
+    return uniqueNamesGenerator({
+      dictionaries: [names],
+      seed: seed,
+    }).toLowerCase();
+  },
+
+  getAWSCredentialsDependingOnEnvironment: () => {
+    if (process.env.NODE_ENV == "production") {
+      return fromInstanceMetadata()();
+    } else {
+      return fromSSO({ profile: "osc-chat" })();
+    }
+  },
 };
